refactor(www): extract langs SVG icons into separate file
Also, rename and move <Editor /> components to components/ui folder.
