+++ apps/www/components/analytics/analytics-bento.tsx
@@ -2,11 +2,13 @@
 import { cn } from "@/lib/utils";
 import { motion } from "framer-motion";
 import { Wand2 } from "lucide-react";
-import { Highlight, type PrismTheme } from "prism-react-renderer";
+import { type PrismTheme } from "prism-react-renderer";
 import { useState } from "react";
 import { PrimaryButton } from "../button";
 import { AnalyticsStars } from "../svg/analytics-stars";
 import { WebAppLight } from "../svg/web-app-light";
+import { CurlIcon, GoIcon, LangIconProps, PythonIcon, RustIcon, TSIcon } from "../svg/lang-icons";
+import { CodeEditor } from "../ui/code-editor";
 
 const theme = {
   plain: {
@@ -75,116 +77,9 @@ export function AnalyticsBento() {
   );
 }
 
-type IconProps = {
-  active: boolean;
-};
-
-const RustIcon: React.FC<IconProps> = ({ active }) => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g opacity={active ? 1 : 0.3}>
-      <path
-        d="M15.7274 16.584C15.2794 16.584 14.8954 16.528 14.5754 16.416C14.2634 16.296 14.0034 16.14 13.7954 15.948C13.5954 15.748 13.4474 15.516 13.3514 15.252C13.2634 14.988 13.2194 14.708 13.2194 14.412V13.752C13.2194 13.472 13.2034 13.244 13.1714 13.068C13.1474 12.892 13.0994 12.756 13.0274 12.66C12.9634 12.564 12.8674 12.5 12.7394 12.468C12.6194 12.436 12.4634 12.42 12.2714 12.42H11.3234V14.688C11.3234 15.056 11.7674 15.24 12.6554 15.24V16.5H7.21939V15.24C7.93139 15.24 8.28739 15.056 8.28739 14.688V9.396C8.28739 9.028 7.93139 8.844 7.21939 8.844V7.584H13.4714C13.8634 7.584 14.2474 7.612 14.6234 7.668C14.9994 7.716 15.3314 7.82 15.6194 7.98C15.9154 8.14 16.1514 8.368 16.3274 8.664C16.5034 8.952 16.5914 9.336 16.5914 9.816C16.5914 10.128 16.5434 10.404 16.4474 10.644C16.3514 10.876 16.2234 11.076 16.0634 11.244C15.9114 11.412 15.7354 11.548 15.5354 11.652C15.3354 11.756 15.1274 11.832 14.9114 11.88C15.3434 11.992 15.6754 12.2 15.9074 12.504C16.1394 12.808 16.2554 13.268 16.2554 13.884V14.712C16.2554 14.888 16.2914 15.02 16.3634 15.108C16.4354 15.188 16.5274 15.228 16.6394 15.228C16.7594 15.228 16.8554 15.188 16.9274 15.108C16.9994 15.02 17.0354 14.888 17.0354 14.712V14.136H17.9114V14.688C17.9114 14.952 17.8674 15.2 17.7794 15.432C17.6994 15.664 17.5674 15.864 17.3834 16.032C17.2074 16.2 16.9834 16.336 16.7114 16.44C16.4394 16.536 16.1114 16.584 15.7274 16.584ZM11.3234 11.22H12.4154C12.6314 11.22 12.8074 11.2 12.9434 11.16C13.0874 11.12 13.1994 11.056 13.2794 10.968C13.3674 10.872 13.4274 10.748 13.4594 10.596C13.4914 10.436 13.5074 10.24 13.5074 10.008C13.5074 9.736 13.4314 9.524 13.2794 9.372C13.1274 9.22 12.8394 9.144 12.4154 9.144H11.3234V11.22Z"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M12 20C16.4183 20 20 16.4183 20 12C20 7.58172 16.4183 4 12 4C7.58172 4 4 7.58172 4 12C4 16.4183 7.58172 20 12 20ZM12 21C16.9706 21 21 16.9706 21 12C21 7.02944 16.9706 3 12 3C7.02944 3 3 7.02944 3 12C3 16.9706 7.02944 21 12 21Z"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-    </g>
-  </svg>
-);
-
-const CurlIcon: React.FC<IconProps> = ({ active }) => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g opacity={active ? 1 : 0.3}>
-      <path
-        d="M18.8202 6.91608C18.6492 6.91608 18.4852 6.84502 18.3643 6.71854C18.2435 6.59205 18.1755 6.4205 18.1755 6.24163C18.1755 6.06275 18.2435 5.8912 18.3643 5.76471C18.4852 5.63823 18.6492 5.56717 18.8202 5.56717C18.9911 5.56717 19.1551 5.63823 19.276 5.76471C19.3969 5.8912 19.4648 6.06275 19.4648 6.24163C19.4648 6.4205 19.3969 6.59205 19.276 6.71854C19.1551 6.84502 18.9911 6.91608 18.8202 6.91608ZM12.4177 18.4394C12.2467 18.4394 12.0827 18.3683 11.9618 18.2418C11.841 18.1154 11.773 17.9438 11.773 17.7649C11.773 17.5861 11.841 17.4145 11.9618 17.288C12.0827 17.1615 12.2467 17.0905 12.4177 17.0905C12.5886 17.0905 12.7526 17.1615 12.8735 17.288C12.9944 17.4145 13.0623 17.5861 13.0623 17.7649C13.0623 17.9438 12.9944 18.1154 12.8735 18.2418C12.7526 18.3683 12.5886 18.4394 12.4177 18.4394ZM18.8202 5.0072C18.5074 5.00772 18.2076 5.13794 17.9864 5.36933C17.7653 5.60071 17.6408 5.91439 17.6403 6.24163C17.6403 6.38685 17.6753 6.52292 17.7197 6.65245L12.1832 16.5796C11.648 16.6973 11.2378 17.1703 11.2378 17.7649C11.2383 18.0922 11.3628 18.4058 11.5839 18.6372C11.8051 18.8686 12.1049 18.9988 12.4177 18.9994C12.7304 18.9988 13.0302 18.8686 13.2514 18.6372C13.4725 18.4058 13.597 18.0922 13.5975 17.7649C13.5975 17.6341 13.5625 17.5033 13.5225 17.3724L19.0871 7.41195C19.6061 7.28111 20 6.82319 20 6.23443C19.9995 5.9072 19.875 5.59352 19.6539 5.36213C19.4327 5.13074 19.1329 5.00052 18.8202 5"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-      <path
-        d="M13.9977 6.91607C13.8267 6.91607 13.6627 6.84501 13.5418 6.71853C13.4209 6.59204 13.353 6.42049 13.353 6.24162C13.353 6.06274 13.4209 5.89119 13.5418 5.7647C13.6627 5.63822 13.8267 5.56716 13.9977 5.56716C14.1686 5.56716 14.3326 5.63822 14.4535 5.7647C14.5744 5.89119 14.6423 6.06274 14.6423 6.24162C14.6423 6.42049 14.5744 6.59204 14.4535 6.71853C14.3326 6.84501 14.1686 6.91607 13.9977 6.91607ZM7.5889 18.44C7.41794 18.44 7.25397 18.369 7.13308 18.2425C7.01219 18.116 6.94427 17.9444 6.94427 17.7656C6.94427 17.5867 7.01219 17.4151 7.13308 17.2887C7.25397 17.1622 7.41794 17.0911 7.5889 17.0911C7.75987 17.0911 7.92383 17.1622 8.04472 17.2887C8.16561 17.4151 8.23353 17.5867 8.23353 17.7656C8.23353 17.9444 8.16561 18.116 8.04472 18.2425C7.92383 18.369 7.75987 18.44 7.5889 18.44ZM13.9977 5.00719C13.3455 5.00719 12.8178 5.55997 12.8178 6.24162C12.8178 6.38684 12.8528 6.52291 12.8972 6.65244L7.35444 16.5796C6.81923 16.6973 6.40907 17.1703 6.40907 17.7649C6.4094 18.0923 6.53379 18.4061 6.75496 18.6376C6.97613 18.8692 7.27603 18.9995 7.5889 19C7.90166 18.9995 8.20147 18.8693 8.42262 18.6379C8.64378 18.4065 8.76824 18.0928 8.76874 17.7656C8.76874 17.6347 8.73372 17.5039 8.69371 17.3731L14.2584 7.41259C14.7773 7.28175 15.1712 6.82383 15.1712 6.23507C15.1691 5.90898 15.0439 5.59697 14.8229 5.367C14.602 5.13702 14.3031 5.0077 13.9914 5.00719M5.17358 8.98065C5.34455 8.98065 5.50851 9.05171 5.6294 9.17819C5.75029 9.30468 5.81821 9.47623 5.81821 9.6551C5.81821 9.83398 5.75029 10.0055 5.6294 10.132C5.50851 10.2585 5.34455 10.3296 5.17358 10.3296C5.00262 10.3296 4.83865 10.2585 4.71776 10.132C4.59687 10.0055 4.52896 9.83398 4.52896 9.6551C4.52896 9.47623 4.59687 9.30468 4.71776 9.17819C4.83865 9.05171 5.00262 8.98065 5.17358 8.98065ZM5.17358 10.8895C5.48634 10.889 5.78615 10.7588 6.00731 10.5274C6.22846 10.296 6.35292 9.98234 6.35342 9.6551C6.35342 9.52427 6.31778 9.39343 6.27839 9.2626C6.12208 8.77197 5.69379 8.41348 5.17296 8.41348C5.0898 8.41348 5.01665 8.44619 4.93787 8.4632C4.40391 8.58095 4 9.05326 4 9.64856C4.0005 9.97579 4.12496 10.2895 4.34611 10.5209C4.56727 10.7523 4.86708 10.8825 5.17984 10.883M4.52896 13.9694C4.52896 13.7905 4.59687 13.619 4.71776 13.4925C4.83865 13.366 5.00262 13.2949 5.17358 13.2949C5.34455 13.2949 5.50851 13.366 5.6294 13.4925C5.75029 13.619 5.81821 13.7905 5.81821 13.9694C5.81821 14.1483 5.75029 14.3198 5.6294 14.4463C5.50851 14.5728 5.34455 14.6438 5.17358 14.6438C5.00262 14.6438 4.83865 14.5728 4.71776 14.4463C4.59687 14.3198 4.52896 14.1483 4.52896 13.9694ZM6.35342 13.9694C6.35342 13.8386 6.31778 13.7077 6.27839 13.5769C6.12208 13.0863 5.69441 12.7278 5.17296 12.7278C5.0898 12.7278 5.01665 12.7605 4.93787 12.7768C4.40391 12.8946 4 13.3675 4 13.9628C4.0005 14.2901 4.12496 14.6038 4.34611 14.8351C4.56727 15.0665 4.86708 15.1968 5.17984 15.1973C5.4926 15.1968 5.7924 15.0665 6.01356 14.8351C6.23471 14.6038 6.35918 14.2901 6.35967 13.9628"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_574_1420"
-          x1="4.15606"
-          y1="2.27462"
-          x2="4.15606"
-          y2="20.9494"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-      </defs>
-    </g>
-  </svg>
-);
-
-const PythonIcon: React.FC<IconProps> = ({ active }) => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      opacity={active ? 1 : 0.3}
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M8.92308 6.34188C8.92308 6.14145 9.10073 5.81724 9.69579 5.51187C10.2601 5.2223 11.0743 5.02564 12 5.02564C12.9257 5.02564 13.7399 5.2223 14.3042 5.51187C14.8993 5.81724 15.0769 6.14145 15.0769 6.34188L15.0768 8.41025V8.58119C15.0768 9.42328 15.0233 10.0118 14.9146 10.4281C14.8088 10.8329 14.6614 11.0364 14.4939 11.1606C14.313 11.2945 14.0407 11.3872 13.5869 11.4368C13.1321 11.4865 12.5656 11.4872 11.829 11.4872H11.803C11.1831 11.4872 10.623 11.4872 10.1514 11.5443C9.6692 11.6028 9.2122 11.7273 8.83772 12.0269C8.4557 12.3325 8.22386 12.7678 8.08632 13.3179C7.96911 13.7868 7.91331 14.3669 7.90039 15.0769H6.34188C6.14145 15.0769 5.81723 14.8993 5.51187 14.3042C5.22229 13.7399 5.02564 12.9257 5.02564 12C5.02564 11.0743 5.22229 10.2601 5.51187 9.6958C5.81723 9.10073 6.14145 8.92308 6.34188 8.92308H8.41026H12V7.89744H8.92308V6.34188ZM6.34188 16.1026H7.89741L7.89744 17.6582C7.89744 18.4489 8.52337 19.0393 9.22754 19.4006C9.96246 19.7777 10.9431 20 12 20C13.0569 20 14.0375 19.7777 14.7724 19.4006C15.4766 19.0393 16.1026 18.4489 16.1026 17.6582V16.1026H17.6581C18.4489 16.1026 19.0393 15.4766 19.4006 14.7724C19.7777 14.0375 20 13.0569 20 12C20 10.9431 19.7777 9.96246 19.4006 9.22754C19.0393 8.52338 18.4489 7.89744 17.6581 7.89744H16.1026V6.34189C16.1026 5.55103 15.4766 4.96071 14.7724 4.59937C14.0375 4.22223 13.0569 4 12 4C10.9431 4 9.96246 4.22223 9.22754 4.59937C8.52337 4.96071 7.89744 5.55103 7.89744 6.34188V7.89744H6.34188C5.55102 7.89744 4.96071 8.52338 4.59937 9.22754C4.22223 9.96246 4 10.9431 4 12C4 13.0569 4.22223 14.0375 4.59937 14.7724C4.96071 15.4766 5.55102 16.1026 6.34188 16.1026ZM12 16.1026H15.0769V17.6582C15.0769 17.8586 14.8993 18.1828 14.3042 18.4881C13.7399 18.7777 12.9257 18.9744 12 18.9744C11.0743 18.9744 10.2601 18.7777 9.69579 18.4881C9.10073 18.1828 8.92308 17.8586 8.92308 17.6582L8.92303 15.5897V15.4188C8.92303 14.5776 8.97644 13.9862 9.08133 13.5667C9.18418 13.1553 9.32627 12.9495 9.47843 12.8278C9.63814 12.7 9.87559 12.6109 10.2748 12.5625C10.678 12.5136 11.1772 12.5128 11.829 12.5128H11.8539C12.56 12.5128 13.1816 12.5128 13.6983 12.4564C14.222 12.3991 14.7082 12.2781 15.1043 11.9847C15.5136 11.6816 15.7615 11.2441 15.9069 10.6873C16.0301 10.2157 16.0865 9.63309 16.0995 8.92308H17.6581C17.8586 8.92308 18.1828 9.10073 18.4881 9.6958C18.7776 10.2601 18.9744 11.0743 18.9744 12C18.9744 12.9257 18.7776 13.7399 18.4881 14.3042C18.1828 14.8993 17.8586 15.0769 17.6581 15.0769H15.5898H12V16.1026ZM10.1806 7.33585C10.5353 7.33585 10.8229 7.04829 10.8229 6.69356C10.8229 6.33884 10.5353 6.05128 10.1806 6.05128C9.82592 6.05128 9.53836 6.33884 9.53836 6.69356C9.53836 7.04829 9.82592 7.33585 10.1806 7.33585ZM13.8227 17.9533C14.1781 17.9533 14.4661 17.6653 14.4661 17.3101C14.4661 16.9547 14.1781 16.6667 13.8227 16.6667C13.4674 16.6667 13.1794 16.9547 13.1794 17.3101C13.1794 17.6653 13.4674 17.9533 13.8227 17.9533Z"
-      fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-    />
-    <defs>
-      <linearGradient
-        id="paint0_linear_574_1420"
-        x1="4.15606"
-        y1="2.27462"
-        x2="4.15606"
-        y2="20.9494"
-        gradientUnits="userSpaceOnUse"
-      >
-        <stop stopColor="white" stopOpacity="0.4" />
-        <stop offset="1" stopColor="white" />
-      </linearGradient>
-    </defs>
-  </svg>
-);
-
-const TSIcon: React.FC<IconProps> = ({ active }) => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <g opacity={active ? 1 : 0.3}>
-      <path
-        d="M12.4559 14.424C12.4559 15.432 13.2239 16.08 14.4159 16.08C15.5839 16.08 16.3439 15.408 16.3439 14.4C16.3439 13.52 15.8319 12.888 14.9199 12.648L14.2079 12.448C13.7919 12.344 13.5679 12.064 13.5679 11.696C13.5679 11.232 13.8799 10.952 14.3999 10.952C14.9359 10.952 15.2559 11.24 15.2559 11.688H16.2559C16.2559 10.704 15.5359 10.08 14.4079 10.08C13.2879 10.08 12.5759 10.72 12.5759 11.712C12.5759 12.568 13.1039 13.2 14.0079 13.44L14.7119 13.632C15.1119 13.736 15.3439 14.04 15.3439 14.432C15.3439 14.904 14.9999 15.208 14.4239 15.208C13.8399 15.208 13.4559 14.896 13.4559 14.424H12.4559Z"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-      <path
-        d="M9.09904 11.064V16H10.107V11.064H11.635V10.16H7.57104V11.064H9.09904Z"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M9.27779 4C8.45652 4 7.80955 3.99999 7.28889 4.04253C6.75771 4.08593 6.31414 4.17609 5.91103 4.38148C5.25247 4.71703 4.71703 5.25247 4.38148 5.91103C4.17609 6.31414 4.08593 6.75771 4.04253 7.28889C3.99999 7.80955 4 8.45652 4 9.27779V14.7222C4 15.5435 3.99999 16.1905 4.04253 16.7111C4.08593 17.2423 4.17609 17.6859 4.38148 18.089C4.71703 18.7475 5.25247 19.283 5.91103 19.6185C6.31414 19.8239 6.75771 19.9141 7.28889 19.9575C7.80953 20 8.45649 20 9.27773 20H14.7222C15.5435 20 16.1905 20 16.7111 19.9575C17.2423 19.9141 17.6859 19.8239 18.089 19.6185C18.7475 19.283 19.283 18.7475 19.6185 18.089C19.8239 17.6859 19.9141 17.2423 19.9575 16.7111C20 16.1905 20 15.5435 20 14.7223V9.27778C20 8.45654 20 7.80953 19.9575 7.28889C19.9141 6.75771 19.8239 6.31414 19.6185 5.91103C19.283 5.25247 18.7475 4.71703 18.089 4.38148C17.6859 4.17609 17.2423 4.08593 16.7111 4.04253C16.1905 3.99999 15.5435 4 14.7222 4H9.27779ZM6.36502 5.27248C6.60366 5.15089 6.90099 5.07756 7.37032 5.03921C7.84549 5.00039 8.45167 5 9.3 5H14.7C15.5483 5 16.1545 5.00039 16.6297 5.03921C17.099 5.07756 17.3963 5.15089 17.635 5.27248C18.1054 5.51217 18.4878 5.89462 18.7275 6.36502C18.8491 6.60366 18.9224 6.90099 18.9608 7.37032C18.9996 7.84549 19 8.45167 19 9.3V14.7C19 15.5483 18.9996 16.1545 18.9608 16.6297C18.9224 17.099 18.8491 17.3963 18.7275 17.635C18.4878 18.1054 18.1054 18.4878 17.635 18.7275C17.3963 18.8491 17.099 18.9224 16.6297 18.9608C16.1545 18.9996 15.5483 19 14.7 19H9.3C8.45167 19 7.84549 18.9996 7.37032 18.9608C6.90099 18.9224 6.60366 18.8491 6.36502 18.7275C5.89462 18.4878 5.51217 18.1054 5.27248 17.635C5.15089 17.3963 5.07756 17.099 5.03921 16.6297C5.00039 16.1545 5 15.5483 5 14.7V9.3C5 8.45167 5.00039 7.84549 5.03921 7.37032C5.07756 6.90099 5.15089 6.60366 5.27248 6.36502C5.51217 5.89462 5.89462 5.51217 6.36502 5.27248Z"
-        fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-      />
-    </g>
-  </svg>
-);
-
-const GoIcon: React.FC<IconProps> = ({ active }) => (
-  <svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path
-      opacity={active ? 1 : 0.3}
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M10.5578 10.1111C10.247 10.1919 9.93194 10.2741 9.56747 10.3666L9.54556 10.3725C9.36723 10.4197 9.34879 10.4244 9.18283 10.2364C8.98432 10.015 8.83829 9.87162 8.55983 9.74107C7.72484 9.33718 6.91624 9.4544 6.16041 9.93662C5.25909 10.51 4.7951 11.3572 4.80843 12.4128C4.82176 13.4553 5.55064 14.3155 6.59802 14.4589C7.49935 14.5761 8.25489 14.2633 8.85165 13.5986C8.94014 13.4922 9.0212 13.3789 9.11142 13.253C9.14261 13.2089 9.17516 13.1636 9.20956 13.1164H6.65106C6.37256 13.1164 6.30619 12.9469 6.39922 12.7255C6.57136 12.3216 6.88987 11.6439 7.07534 11.305C7.11506 11.2269 7.20778 11.0966 7.40665 11.0966H11.6727C11.8642 10.4997 12.1753 9.93579 12.5899 9.40218C13.5576 8.15134 14.7241 7.49939 16.3015 7.22606C17.6538 6.99134 18.9263 7.12161 20.0795 7.89051C21.127 8.5944 21.7763 9.54551 21.9486 10.7966C22.174 12.5561 21.657 13.9897 20.4244 15.2147C19.5494 16.0878 18.4757 16.635 17.2428 16.8828C17.0074 16.9256 16.7723 16.9458 16.5409 16.9658C16.4203 16.9761 16.3004 16.9867 16.1825 17C14.9759 16.9739 13.8759 16.635 12.9478 15.8533C12.2952 15.2986 11.8454 14.6169 11.6222 13.8233C11.4654 14.1342 11.2789 14.4299 11.0652 14.7064C10.1109 15.9444 8.86465 16.7133 7.28748 16.9219C5.98826 17.0914 4.78177 16.8436 3.72135 16.0617C2.74034 15.3319 2.18361 14.3675 2.03782 13.1686C1.86539 11.748 2.28967 10.4711 3.16462 9.35023C4.10593 8.13829 5.35184 7.36939 6.87625 7.09578C8.12247 6.87411 9.31537 7.01745 10.3893 7.73411C11.0919 8.19023 11.5956 8.81579 11.927 9.57163C12.0065 9.68913 11.9535 9.75413 11.7944 9.7933C11.3815 9.89679 10.9694 10.0029 10.5578 10.1111ZM19.328 11.5958C19.3307 11.6394 19.3336 11.6855 19.3373 11.735C19.271 12.8558 18.7009 13.69 17.6538 14.2242C16.9512 14.5761 16.2221 14.615 15.493 14.3022C14.5386 13.8855 14.0349 12.8558 14.2734 11.8394C14.5649 10.6144 15.3605 9.84524 16.5933 9.57163C17.8526 9.28496 19.0588 10.0147 19.2976 11.305C19.3164 11.3972 19.3217 11.4897 19.328 11.5958Z"
-      fill={active ? "url(#paint0_linear_574_1420)" : "white"}
-    />
-  </svg>
-);
-
 type Language = {
   name: string;
-  Icon: React.FC<IconProps>;
+  Icon: React.FC<LangIconProps>;
   codeBlock: string;
 };
 
@@ -253,7 +148,6 @@ func main() {
 		// handle response
 	}
 }
-
 `;
 
 const rustCodeBlock = `use unkey::models::GetKeyRequest;
@@ -294,7 +188,7 @@ function AnalyticsApiView() {
       <div className="w-full analytics-background-gradient bg-black bg-opacity-02 lg:rounded-3xl xxl:mr-10 overflow-x-hidden overflow-y-hidden border-white/10 border border-b-0 border-l-0  border-r-0 flex-col md:flex-row relative rounded-tl-3xl h-[600px] xl:h-[576px] flex">
         <LanguageSwitcher languages={Object.values(languagesList)} currentLanguage={language} setLanguage={setLanguage} />
         <div className="flex pt-4 pb-8 pl-8 font-mono text-xs text-white sm:text-sm">
-          <Editor theme={theme} codeBlock={languagesList[language].codeBlock} language="tsx" />
+          <CodeEditor theme={theme} codeBlock={languagesList[language].codeBlock} language="tsx" />
           <button
               type="button"
               aria-label="Copy code"
@@ -611,7 +505,7 @@ function AnalyticsWebAppView() {
       whileInView="visible"
       className="w-full overflow-x-hidden"
     >
-      <div className="w-full bg-[#000000] lg:w-[1220px] analytics-background-gradient lg:rounded-2xl relative flex-wrap md:flex-nowrap cursor-default select-none bg-opacity-02 xxl:mr-10 overflow-x-hidden overflow-y-hidden border-white/10 border border-b-0 border-l-0  border-r-0 flex-col md:flex-row relative rounded-tl-3xl h-[600px] xl:h-[576px] flex">
+      <div className="w-full bg-[#000000] lg:w-[1220px] analytics-background-gradient lg:rounded-2xl flex-wrap md:flex-nowrap cursor-default select-none bg-opacity-02 xxl:mr-10 overflow-x-hidden overflow-y-hidden border-white/10 border border-b-0 border-l-0  border-r-0 flex-col md:flex-row relative rounded-tl-3xl h-[600px] xl:h-[576px] flex">
         <WebAppLight className="absolute top-[-100px] left-[40px]" />
         <div className="flex flex-col rounded-2xl w-[216px] h-full text-white/20  text-xs pt-6 px-4 md:border-r md:border-white/5">
           <div className="flex justify-between items-center w-[160px]">
@@ -926,140 +820,6 @@ function AnalyticsWebAppView() {
         </div>
       </div>
     </motion.div>
-  );
-}
-
-export function Editor({
-  codeBlock,
-  language,
-  theme,
-}: { codeBlock: string; language: string; theme?: PrismTheme }) {
-  return (
-    <Highlight theme={theme} code={codeBlock} language={language}>
-      {({ tokens, getLineProps, getTokenProps }) => {
-        const amountLines = tokens.length;
-        const gutterPadLength = Math.max(String(amountLines).length, 2);
-        return (
-          <pre
-            key={codeBlock} // Use codeBlock as a key to trigger animations on change
-            className="leading-6"
-          >
-            {tokens.map((line, i) => {
-              const lineNumber = i + 1;
-              const paddedLineGutter = String(lineNumber).padStart(gutterPadLength, " ");
-              return (
-                // biome-ignore lint/suspicious/noArrayIndexKey: I got nothing better right now
-                <div key={`${line}-${i}`} {...getLineProps({ line })}>
-                  <span className="select-none line-number">{paddedLineGutter}</span>
-                  {line.map((token, key) => (
-                    <span key={`${key}-${token}`} {...getTokenProps({ token })} />
-                  ))}
-                </div>
-              );
-            })}
-          </pre>
-        );
-      }}
-    </Highlight>
-  );
-}
-
-export function TerminalIcon({ className }: { className?: string }) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      className={className}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M17.9864 5.36932C18.2076 5.13794 18.5073 5.00771 18.82 5.0072V5C19.1328 5.00051 19.4326 5.13074 19.6538 5.36214C19.875 5.59352 19.9993 5.90721 20 6.23444C20 6.82321 19.606 7.28113 19.0871 7.41197L13.5224 17.3725C13.5623 17.5034 13.5973 17.6342 13.5973 17.7651C13.5968 18.0923 13.4725 18.406 13.2513 18.6374C13.0301 18.8688 12.7303 18.999 12.4175 18.9995C12.1049 18.999 11.8048 18.8688 11.5839 18.6374C11.3627 18.406 11.2382 18.0923 11.2377 17.7651C11.2377 17.1704 11.6478 16.6974 12.183 16.5797L17.7195 6.65246C17.6752 6.52294 17.6402 6.38687 17.6402 6.24164C17.6409 5.9144 17.7652 5.60072 17.9864 5.36932ZM18.3643 6.71855C18.4852 6.84504 18.649 6.91611 18.82 6.91611C18.991 6.91611 19.155 6.84504 19.2758 6.71855C19.3967 6.59207 19.4646 6.42052 19.4646 6.24164C19.4646 6.06275 19.3967 5.89121 19.2758 5.76472C19.155 5.63824 18.991 5.56718 18.82 5.56718C18.649 5.56718 18.4852 5.63824 18.3643 5.76472C18.2435 5.89121 18.1754 6.06275 18.1754 6.24164C18.1754 6.42052 18.2435 6.59207 18.3643 6.71855ZM11.9616 18.242C12.0827 18.3685 12.2465 18.4396 12.4175 18.4396C12.5885 18.4396 12.7525 18.3685 12.8734 18.242C12.9942 18.1155 13.0621 17.9439 13.0621 17.7651C13.0621 17.5862 12.9942 17.4146 12.8734 17.2882C12.7525 17.1617 12.5885 17.0906 12.4175 17.0906C12.2465 17.0906 12.0827 17.1617 11.9616 17.2882C11.8408 17.4146 11.7729 17.5862 11.7729 17.7651C11.7729 17.9439 11.8408 18.1155 11.9616 18.242Z"
-        fill="url(#paint0_linear_840_1992)"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M12.8178 6.24152C12.8178 5.56123 13.3435 5.00929 13.9937 5.00708C14.3044 5.00825 14.6024 5.13751 14.8231 5.36688C15.044 5.59687 15.169 5.90888 15.1714 6.23497C15.1714 6.82374 14.7773 7.28166 14.2585 7.4125L8.69378 17.3732C8.73367 17.5039 8.76867 17.6347 8.76867 17.7656C8.7682 18.0929 8.64385 18.4065 8.42268 18.6379C8.20151 18.8693 7.90172 18.9995 7.58887 19C7.27601 18.9995 6.97622 18.8693 6.75505 18.6377C6.53388 18.4061 6.4093 18.0923 6.40906 17.765C6.40906 17.1703 6.81921 16.6973 7.3544 16.5795L12.8972 6.65235C12.8528 6.52282 12.8178 6.38675 12.8178 6.24152ZM13.5418 6.71844C13.6629 6.84493 13.8266 6.91598 13.9976 6.91598C14.1686 6.91598 14.3327 6.84493 14.4535 6.71844C14.5744 6.59195 14.6422 6.4204 14.6422 6.24152C14.6422 6.06264 14.5744 5.89108 14.4535 5.7646C14.3327 5.63811 14.1686 5.56705 13.9976 5.56705C13.8266 5.56705 13.6629 5.63811 13.5418 5.7646C13.4209 5.89108 13.353 6.06264 13.353 6.24152C13.353 6.4204 13.4209 6.59195 13.5418 6.71844ZM7.133 18.2425C7.25408 18.369 7.41786 18.4401 7.58887 18.4401C7.75988 18.4401 7.92389 18.369 8.04474 18.2425C8.16559 18.116 8.23348 17.9446 8.23348 17.7656C8.23348 17.5867 8.16559 17.4151 8.04474 17.2887C7.92389 17.1622 7.75988 17.0912 7.58887 17.0912C7.41786 17.0912 7.25408 17.1622 7.133 17.2887C7.01215 17.4151 6.94426 17.5867 6.94426 17.7656C6.94426 17.9446 7.01215 18.116 7.133 18.2425Z"
-        fill="url(#paint1_linear_840_1992)"
-      />
-      <path
-        d="M13.9937 5.00708C13.9951 5.00708 13.9962 5.00707 13.9976 5.00707H13.9913C13.992 5.00707 13.993 5.00708 13.9937 5.00708Z"
-        fill="url(#paint2_linear_840_1992)"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.00732 10.5274C5.78615 10.7588 5.48636 10.889 5.1735 10.8895L5.1798 10.883C4.86718 10.8824 4.56715 10.7522 4.34622 10.5208C4.12505 10.2894 4.00047 9.97574 4 9.64851C4 9.0532 4.40384 8.58088 4.93787 8.46312C4.96377 8.45755 4.98896 8.45029 5.01416 8.44305C5.06572 8.42819 5.11704 8.41341 5.17304 8.41341C5.69376 8.41341 6.1221 8.7719 6.27842 9.26253C6.31784 9.39337 6.35354 9.52422 6.35354 9.65506C6.35284 9.98228 6.22849 10.296 6.00732 10.5274ZM5.62937 9.17814C5.50852 9.05165 5.34451 8.98059 5.1735 8.98059C5.00273 8.98059 4.83872 9.05165 4.71787 9.17814C4.59678 9.30461 4.52889 9.47617 4.52889 9.65506C4.52889 9.83392 4.59678 10.0055 4.71787 10.132C4.83872 10.2585 5.00273 10.3295 5.1735 10.3295C5.34451 10.3295 5.50852 10.2585 5.62937 10.132C5.75022 10.0055 5.81811 9.83392 5.81811 9.65506C5.81811 9.47617 5.75022 9.30461 5.62937 9.17814Z"
-        fill="url(#paint3_linear_840_1992)"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M6.27842 13.5769C6.31784 13.7077 6.35354 13.8386 6.35354 13.9694L6.3596 13.9628C6.35914 14.2901 6.23479 14.6038 6.01362 14.8352C5.79245 15.0666 5.49266 15.1968 5.1798 15.1973C4.86718 15.1968 4.56715 15.0666 4.34622 14.8352C4.12505 14.6038 4.00047 14.2901 4 13.9628C4 13.3675 4.40384 12.8946 4.93787 12.7768C4.96423 12.7713 4.99013 12.764 5.01579 12.7567C5.06688 12.7422 5.11774 12.7277 5.17304 12.7277C5.69446 12.7277 6.1221 13.0862 6.27842 13.5769ZM4.71787 13.4925C4.59678 13.619 4.52889 13.7905 4.52889 13.9694C4.52889 14.1483 4.59678 14.3198 4.71787 14.4463C4.83872 14.5728 5.00273 14.6439 5.1735 14.6439C5.34451 14.6439 5.50852 14.5728 5.62937 14.4463C5.75022 14.3198 5.81811 14.1483 5.81811 13.9694C5.81811 13.7905 5.75022 13.619 5.62937 13.4925C5.50852 13.366 5.34451 13.2949 5.1735 13.2949C5.00273 13.2949 4.83872 13.366 4.71787 13.4925Z"
-        fill="url(#paint4_linear_840_1992)"
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_840_1992"
-          x1="4.15606"
-          y1="3.49029"
-          x2="4.15606"
-          y2="19.8307"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_840_1992"
-          x1="4.15606"
-          y1="3.49029"
-          x2="4.15606"
-          y2="19.8307"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-        <linearGradient
-          id="paint2_linear_840_1992"
-          x1="4.15606"
-          y1="3.49029"
-          x2="4.15606"
-          y2="19.8307"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-        <linearGradient
-          id="paint3_linear_840_1992"
-          x1="4.15606"
-          y1="3.49029"
-          x2="4.15606"
-          y2="19.8307"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-        <linearGradient
-          id="paint4_linear_840_1992"
-          x1="4.15606"
-          y1="3.49029"
-          x2="4.15606"
-          y2="19.8307"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="white" stopOpacity="0.4" />
-          <stop offset="1" stopColor="white" />
-        </linearGradient>
-      </defs>
-    </svg>
   );
 }
